moved the user feed interface into the interface-types folder for consistent file structure
